requirements.txt
streamlit
transformers
torch
sentence-transformers
faiss-cpu
pypdf
numpy
llama-cpp-python